+++ src/components/Form/FormItem/FormItem.styled.tsx
@@ -1,7 +1,8 @@
-import styled, { css } from "styled-components";
+import styled from "styled-components";
 
 import { FormItemProps } from "../FormItem/FormItem";
 import { InputFieldStyled } from "../InputField/InputField.styled";
+import { SelectWrapperStyled } from "../Select/Select.styled";
 import { TextAreaStyled } from "../TextArea/TextArea.styled";
 
 
@@ -41,32 +42,60 @@ export const FormItemStyled = styled.div`
 
         span:not(:last-of-type) {
             margin-right: 10px;
-
-            /* &::after {
-                content: "\\0000a0\\2022\\0000a0";
-                margin-left: 5px;
-            } */
         }
     }
 
-
-    ${InputFieldStyled}, ${TextAreaStyled} {
-        &:focus ~ label { font-weight : bold; }
-
-        &:required ~ label::before {
-            position  : absolute;
-            top       : -4px;
-            bottom    : 0;
-            left      : -12px;
-            margin    : auto;
-            content   : "*";
-            font-size : 120%;
-            color     : ${(props: FormItemProps) => props.theme.inputField.required.text};
+    &[required] {
+        & > div, // For RadioGroup
+        & > ${InputFieldStyled}, 
+        & > ${TextAreaStyled}, 
+        & > ${SelectWrapperStyled} {
+            & ~ label::before {
+                position  : absolute;
+                top       : -4px;
+                bottom    : 0;
+                left      : -12px;
+                margin    : auto;
+                content   : "*";
+                font-size : 120%;
+                color     : ${(props: FormItemProps) => props.theme.inputField.required.text};
+            }
         }
+    }
+
+    input[type="radio"]:disabled + label,
+    input[type="checkbox"]:disabled + label {
+        pointer-events : none;
+        user-select    : none;
+        cursor         : default;
+        opacity        : 0.36;
+    }
     
-        &:invalid:not(:focus):not(:placeholder-shown) ~ .info-section span.error-text {
-            display: block;
-            color : ${(props: FormItemProps) => props.theme.inputField.isInvalid.errorText};
+    ${InputFieldStyled}, ${TextAreaStyled}, ${SelectWrapperStyled} {
+        &:read-only {
+            background-color : ${(props: FormItemProps) => props.theme.inputField.isReadOnly.bg};
+            color            : ${(props: FormItemProps) => props.theme.inputField.default.text};
+
+            &:focus { border : 2px solid ${(props: FormItemProps) => props.theme.inputField.default.text}; }
+        }
+
+        &:disabled, &[disabled] select {
+            background-color : ${(props: FormItemProps) => props.theme.inputField.isReadOnly.bg};
+            border-color     : ${(props: FormItemProps) => props.theme.inputField.isReadOnly.border};
+            color            : ${(props: FormItemProps) => props.theme.inputField.isReadOnly.text};
+            
+            & + label {
+                color : ${(props: FormItemProps) => props.theme.inputField.isReadOnly.label};
+            }
+        }
+
+        &:not(:disabled) {
+            &:focus ~ label { font-weight : bold; }
+            
+            &:invalid:not(:focus):not(:placeholder-shown) ~ .info-section span.error-text {
+                display: block;
+                color : ${(props: FormItemProps) => props.theme.inputField.isInvalid.errorText};
+            }
         }
     }
 `;